extract cmds and directory diffing into their own classes
