+++ lib/semgrep.rb
@@ -2,6 +2,7 @@ require 'open3'
 require 'json'
 require_relative 'semgrep/finding'
 require_relative 'semgrep/runner'
+require_relative 'semgrep/directory_differ'
 
 module Semgrep
 end
