+++ lib/whiskers.rb
@@ -4,6 +4,9 @@ require_relative 'whiskers/dependency_change'
 require_relative 'whiskers/gemfile_lock_parser'
 require_relative 'whiskers/gemfile_lock_diff'
 require_relative 'whiskers/gem_version_comparer'
+require_relative 'whiskers/commands/gem_diff'
+require_relative 'whiskers/commands/lockfile_diff'
+require_relative 'whiskers/commands/gem_diff_bulk'
 require_relative 'whiskers/cli'
 require_relative 'semgrep'
